Update mozjs to 87

Depends on: https://github.com/servo/rust-mozjs/pull/540

+++ components/script/dom/imagedata.rs
@@ -113,7 +113,8 @@ impl ImageData {
         let len = width * height * 4;
         let cx = global.get_cx();
         rooted!(in (*cx) let mut array = ptr::null_mut::<JSObject>());
-        Uint8ClampedArray::create(*cx, CreateWith::Length(len), array.handle_mut()).unwrap();
+        Uint8ClampedArray::create(*cx, CreateWith::Length(len as usize), array.handle_mut())
+            .unwrap();
         (*imagedata).data.set(array.get());
 
         Ok(reflect_dom_object(imagedata, global))
